Project4/requirements.txt
langgraph==0.2.28
langchain==0.3.7
langchain-openai==0.2.6
openai==1.58.1
streamlit==1.33.0
python-dotenv==1.0.1
